Moved QR handling into a separate function so it could work with the FDC callback. App can now determine whether or not a QR code exists and then decide what flow the user goes down based on that
