+++ app/app/src/main/java/com/example/scavengingscannables/ScannerActivity.java
@@ -40,33 +40,14 @@ import java.util.Locale;
 
 public class ScannerActivity extends AppCompatActivity {
     private CodeScanner mCodeScanner;
-
     private FirestoreDatabaseController fdc = new FirestoreDatabaseController();
-
     private ScoringSystem scrsys = new ScoringSystem();
-
-    private NamingSystem namsys = new NamingSystem();
-
-    private boolean storePhoto;
-
-    private boolean storeLocation;
-
-    int CAMERA_REQUEST_CODE = 102;
-
-    double latitude;
-
-    double longitude;
-
-    FusedLocationProviderClient flpc;
-
-    boolean codeExists;
+    private QRCodeHandler qrch;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.scanner);
-
-        flpc = LocationServices.getFusedLocationProviderClient(ScannerActivity.this);
 
         CodeScannerView scannerView = findViewById(R.id.scanner_view);
         mCodeScanner = new CodeScanner(this, scannerView);
@@ -85,17 +66,12 @@ public class ScannerActivity extends AppCompatActivity {
                         // Generate a score for the hash
                         int score = scrsys.generateScore(sha256hex);
 
+                        qrch = new QRCodeHandler(ScannerActivity.this, sha256hex, score, fdc);
+
                         // Tell the user what the score of the QR code they scanned was
                         Toast.makeText(ScannerActivity.this, "Your score is: " + score,Toast.LENGTH_SHORT).show();
 
-//                        codeExists = fdc.CheckQRIDExists(sha256hex, );
-
-                        if (codeExists) {
-                            scanNewQRCode(sha256hex, score);
-                        }
-                        else {
-                            scanExistingQRCode();
-                        }
+                        fdc.CheckQRIDExists(sha256hex, qrch);
                     }
                 });
             }
@@ -107,117 +83,6 @@ public class ScannerActivity extends AppCompatActivity {
                 mCodeScanner.startPreview();
             }
         });
-    }
-
-    private void scanNewQRCode(String hash, int score) {
-        // Ask the user if they want to store an image of the object they just scanned
-        // Then, whether the user wants to store an image or not, we ask if they want to store the location of the object they just scanned
-        askForPhoto();
-
-        // Generate a name for the hash
-        String hashedName = namsys.generateName(hash);
-
-        // Create HashMap for comments, ArrayList for owners, and an ArrayList for locations
-        HashMap<String, String> comments = new HashMap<String, String>();
-        ArrayList<String> ownedBy = new ArrayList<>();
-
-        // Store location based on the user's choices
-        ArrayList<Double> qrLocation = new ArrayList<>();
-        if (storeLocation) {
-            qrLocation.add(latitude);
-            qrLocation.add(longitude);
-        }
-
-        // Create a QR code using the data we've generated
-        QrCode newCode = new QrCode(hash, Integer.toString(score), hashedName,  comments, ownedBy, qrLocation);
-
-        // Save the new QR code to the database
-        fdc.SaveQRCodeByID(newCode);
-    }
-
-    private void scanExistingQRCode() {
-       // Get QR code using its id
-       // Add the current user to its ownedBy list
-    }
-
-    private void askLocationPermissions() {
-        // Here we ask the user if they want to store the object's location
-        new AlertDialog.Builder(ScannerActivity.this)
-                .setTitle("Do you want to store this code's location?")
-
-                // If the user wants to store the location, we set the "storeLocation" flag to true
-                .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        storeLocation = true;
-                        if (ActivityCompat.checkSelfPermission(ScannerActivity.this, Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
-//                           getLocation();
-                            flpc.getLastLocation().addOnCompleteListener(new OnCompleteListener<Location>() {
-                                @Override
-                                public void onComplete(@NonNull Task<Location> task) {
-                                    Location location = task.getResult();
-                                    if (location != null) {
-                                        try {
-                                            // Initialize geoCoder
-                                            Geocoder geocoder = new Geocoder(ScannerActivity.this, Locale.getDefault());
-                                            // Initialize address list
-                                            List<Address> addresses = geocoder.getFromLocation(location.getLatitude(), location.getLongitude(), 1);
-                                            Toast.makeText(ScannerActivity.this, addresses.get(0).getLatitude() + ", " + addresses.get(0).getLongitude(), Toast.LENGTH_LONG).show();
-                                            // Store latitude and longitude
-                                            latitude = addresses.get(0).getLatitude();
-                                            longitude = addresses.get(0).getLongitude();
-
-                                        } catch (IOException e) {
-                                            e.printStackTrace();
-                                        }
-                                    }
-                                }
-                            });
-                        }
-                        else {
-                            ActivityCompat.requestPermissions(ScannerActivity.this, new String[]{Manifest.permission.ACCESS_FINE_LOCATION}, 44);
-                        }
-                    }
-                })
-
-                // If the user does not want to store the location, we set the "storeLocation" flag to false
-                .setNegativeButton("No", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        storeLocation = false;
-                        dialog.dismiss();
-                    }
-                })
-                .create().show();
-    }
-    private void askForPhoto() {
-        new AlertDialog.Builder(ScannerActivity.this)
-                .setTitle("Do you want to store an image of the object you just scanned?")
-
-                // If the user decides they want to store and image, we will launch their phone's camera application
-                // We will also set the "storeCode" flag to true
-                // Then we ask if they want to store the location of where they scanned the code
-                .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        storePhoto= true;
-                        // Launch camera activity
-                        Intent myIntent = new Intent(ScannerActivity.this, CameraActivity.class);
-//                      myIntent.putExtra("key", value); //Optional parameters
-                        ScannerActivity.this.startActivity(myIntent);
-                        askLocationPermissions();
-                    }
-                })
-                // If the user decides they do not want to store the image, we will go straight to asking them if they want to store the location of where they scanned the image
-                .setNegativeButton("No", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        storePhoto = false;
-                        dialog.dismiss();
-                        askLocationPermissions();
-                    }
-                })
-                .create().show();
     }
 
     @Override
